refactor: Make API relative instead of absolute
The API URI is currently set absolutely and requires a host name be
provided via environmental variables, this should instead be relative to
the host.

Update .env and Dockerfile to remove the environmental variable/

Update apiService to create the Axios instance in the constructor.
Update ProfileService to extend apiService and pass in `/trainee/api` as
baseUrl.

TISNEW-3848