fix: Add pre-prop specialty tie breaker to FormRs
The pre-populated specialty for Form R Part A/B is based on the lastest
programme's latest curricula. In the case of a tie, i.e. two curricula
starting on the same day the specialty pre-populated is random.
Apply an alphabetical sort as a tie breaker, the first specialty
alphabetically should be the on populated.

TIS21-1397